Refactoring tests (#9320)
# Conflicts:
#	testing/tests/DevExpress.ui.widgets.scheduler/integration.appointments.tests.js
